wpsc
This project provides a command line interface for the Web Power Switch
by Digital Loggers. The WPS lacks a straight forward scheduling
interface and has limited scripting, so using this script the user can
add cron jobs or similar scheduled outlet switches.  It requires a *nix
system with libcurl to run. It was coded on Mac OS X 10.9 and has been
tested on Ubuntu 12.04 LTS and Raspbian.